Tests: Refactor test environment bootstrapping
