+++ tests/conftest.py
@@ -3,6 +3,25 @@ from pathlib import Path
 
 import pytest
 
+from influxio.adapter import InfluxDbApiAdapter, SqlAlchemyAdapter
+
+CRATEDB_URL = "crate://crate@localhost:4200/testdrive/basic"
+INFLUXDB_API_URL = "http://example:token@localhost:8086/testdrive/basic"
+
+
+@pytest.fixture
+def influxdb() -> InfluxDbApiAdapter:
+    adapter = InfluxDbApiAdapter.from_url(INFLUXDB_API_URL)
+    adapter.delete_bucket(missing_ok=True)
+    return adapter
+
+
+@pytest.fixture
+def cratedb() -> SqlAlchemyAdapter:
+    adapter = SqlAlchemyAdapter.from_url(CRATEDB_URL)
+    adapter.run_sql("DROP TABLE IF EXISTS basic")
+    return adapter
+
 
 @pytest.fixture
 def line_protocol_stream_basic():
